Forkgc fixes

+++ src/gc.c
@@ -8,40 +8,22 @@
 #include <assert.h>
 
 static void BlockClients_push(BlockClients* ctx, RedisModuleBlockedClient* bClient) {
-  pthread_mutex_lock(&ctx->lock);
   BlockClient* bc = rm_calloc(1, sizeof(BlockClient));
   bc->bClient = bClient;
-
-  if (ctx->head == NULL) {
-    ctx->head = ctx->tail = bc;
-    pthread_mutex_unlock(&ctx->lock);
-    return;
-  }
-
-  bc->next = ctx->head;
-  ctx->head->prev = bc;
-  ctx->head = bc;
+  pthread_mutex_lock(&ctx->lock);
+  dllist_prepend(&ctx->clients, &bc->llnode);
   pthread_mutex_unlock(&ctx->lock);
 }
 
 static RedisModuleBlockedClient* BlockClients_pop(BlockClients* ctx) {
   pthread_mutex_lock(&ctx->lock);
-  BlockClient* bc = ctx->tail;
-  if (!bc) {
-    pthread_mutex_unlock(&ctx->lock);
-    return NULL;
+  RedisModuleBlockedClient* ret = NULL;
+  DLLIST_node* nn = dllist_pop_tail(&ctx->clients);
+  if (nn) {
+    BlockClient* bc = DLLIST_ITEM(nn, BlockClient, llnode);
+    ret = bc->bClient;
+    rm_free(bc);
   }
-
-  ctx->tail = bc->prev;
-  if (ctx->tail) {
-    ctx->tail->next = NULL;
-  } else {
-    ctx->head = NULL;
-  }
-
-  RedisModuleBlockedClient* ret = bc->bClient;
-  rm_free(bc);
-
   pthread_mutex_unlock(&ctx->lock);
   return ret;
 }
@@ -50,9 +32,10 @@ GCContext* GCContext_CreateGCFromSpec(IndexSpec* sp, float initialHZ, uint64_t u
                                       uint32_t gcPolicy) {
   GCContext* ret = rm_calloc(1, sizeof(GCContext));
   pthread_mutex_init(&ret->bClients.lock, NULL);
+  dllist_init(&ret->bClients.clients);
   switch (gcPolicy) {
     case GCPolicy_Fork:
-      ret->gcCtx = NewForkGCFromSpec(sp, uniqueId, &ret->callbacks);
+      ret->gcCtx = FGC_NewFromSpec(sp, uniqueId, &ret->callbacks);
       break;
     case GCPolicy_Default:
     default:
@@ -65,10 +48,11 @@ GCContext* GCContext_CreateGCFromSpec(IndexSpec* sp, float initialHZ, uint64_t u
 
 GCContext* GCContext_CreateGC(RedisModuleString* keyName, float initialHZ, uint64_t uniqueId) {
   GCContext* ret = rm_calloc(1, sizeof(GCContext));
+  dllist_init(&ret->bClients.clients);
   pthread_mutex_init(&ret->bClients.lock, NULL);
   switch (RSGlobalConfig.gcPolicy) {
     case GCPolicy_Fork:
-      ret->gcCtx = NewForkGC(keyName, uniqueId, &ret->callbacks);
+      ret->gcCtx = FGC_New(keyName, uniqueId, &ret->callbacks);
       break;
     case GCPolicy_Default:
     default:
@@ -93,9 +77,10 @@ static int GCContext_PeriodicCallback(RedisModuleCtx* ctx, void* privdata) {
   if (bClient) {
     RedisModule_UnblockClient(bClient, NULL);
   }
-
-  RMUtilTimer_SetInterval(gc->timer, gc->callbacks.getInterval(gc->gcCtx));
-
+  if (gc->timer) {
+    // Timer could've been deleted..
+    RMUtilTimer_SetInterval(gc->timer, gc->callbacks.getInterval(gc->gcCtx));
+  }
   return ret;
 }
 
@@ -111,9 +96,10 @@ void GCContext_Start(GCContext* gc) {
 }
 
 void GCContext_Stop(GCContext* gc) {
-  if (gc->timer) {
-    RMUtilTimer_Terminate(gc->timer);
+  if (gc->callbacks.kill) {
+    gc->callbacks.kill(gc->gcCtx);
   }
+  RMUtilTimer_Terminate(gc->timer);
 }
 
 void GCContext_RenderStats(GCContext* gc, RedisModuleCtx* ctx) {
@@ -121,7 +107,9 @@ void GCContext_RenderStats(GCContext* gc, RedisModuleCtx* ctx) {
 }
 
 void GCContext_OnDelete(GCContext* gc) {
-  gc->callbacks.onDelete(gc->gcCtx);
+  if (gc->callbacks.onDelete) {
+    gc->callbacks.onDelete(gc->gcCtx);
+  }
 }
 
 void GCContext_ForceInvoke(GCContext* gc, RedisModuleBlockedClient* bc) {
